backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
pydantic==2.11.7
pydantic-settings
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
protobuf==6.31.1
python-multipart==0.0.9
aiosqlite==0.20.0
greenlet==3.0.3
grpcio-tools==1.75.0
email-validator
pydantic_validation_decorator
loguru
httpx
